fix(security): address review findings on gosec G115 PR
Two independent adversarial reviewers of this PR found and confirmed
real defects among the 199 suppressions and missing test coverage on
the 12 fixes -- addressing all of them:

- server/services/github_service.go: PRComment.id was suppressed as
  "far below int32 range" but GitHub's own comment IDs already exceed
  int32 max in production (observed values over 5 billion). Widened
  proto session.v1.PRComment.id to int64 and updated the one Go call
  site and its frontend consumers (VcsWidgetComments.tsx's React key
  and comment-link href, plus two test files' mock literals) for the
  resulting bigint field.
- server/services/connectrpc_websocket.go: marshalProtoEnvelope wrote
  the same 4-byte length prefix as server/protocol/envelope.go's
  CreateEnvelope with no overflow guard; added one, returning an error
  (this function already has an error return, so no panic needed).
- session/instance_tmux.go and session/tmux/tmux.go each independently
  defined an identical clampWinsizeDim helper (added by different
  parallel batches fixing this same PR's findings). Exported
  tmux.ClampWinsizeDim and removed the session-package duplicate, used
  by both callers now (instance_tmux.go and native_process_manager.go).
- server/services/rules_service.go and backlog_service_query.go: 28
  #nosec G115 comments had no reason after the suppression, unlike
  every other site in this PR -- filled in the specific bounding
  rationale for each.
- Added regression tests for all 9 real containment/validation fixes
  that shipped with none (rules_service.go's YAML priority validation,
  session_service.go's ForkAtMessage check, remote_service.go's port
  validation, envelope.go's CreateEnvelope overflow panic,
  tmux.ClampWinsizeDim, tymux's clampResizeDim and
  parseScrollbackOffset, scrollback/manager.go's pagination clamp) --
  the same gap PR #679's review found and required a follow-up for.
- session/scrollback/buffer.go: writing the regression test for
  GetScrollbackBefore surfaced a real bug the manager.go G115 fix
  introduced -- clamping an out-of-range beforeSeq to math.MaxInt
  (correctly avoiding the uint64->int sign-flip) then passing that
  straight through as a slice pre-allocation capacity, which panics
  with "makeslice: cap out of range". Fixed at the root in
  CircularBuffer.GetRange: cap the pre-allocation to the buffer's
  actual entry count (it can never return more than that regardless of
  what limit any caller passes), which fixes every caller, not just
  this one path.

Verified: go build ./... clean; gosec G115 still 0, no new findings
(241 total, unchanged); go test ./... passes, all 66 packages;
web-app tsc --noEmit and jest (93 tests) both clean.

+++ server/services/github_service.go
@@ -118,9 +118,11 @@ func (gs *GitHubService) GetPRComments(
 	protoComments := make([]*sessionv1.PRComment, 0, len(comments))
 	for _, comment := range comments {
 		protoComment := &sessionv1.PRComment{
-			// #nosec G115 -- comment.ID is a GitHub comment ID (gh CLI JSON),
-			// far below int32 range for the foreseeable future.
-			Id:        int32(comment.ID),
+			// GitHub comment IDs already exceed int32 range in practice
+			// (observed values over 5 billion as of 2026-09) -- PRComment.id
+			// is int64 in the proto for exactly this reason, so no G115
+			// conversion happens here.
+			Id:        int64(comment.ID),
 			Author:    comment.Author,
 			Body:      comment.Body,
 			CreatedAt: timestamppb.New(comment.CreatedAt),
